requirements.txt
streamlit
openai
scikit-learn
pymupdf
python-docx
python-pptx
